Move add_reader_themes_request_filter into Trait
